stash: avoid unnecessary reset_tree() call
In 45c1389c31b4 (stash: convert apply to builtin, 2018-12-20), we
introduced code that is the equivalent of `git write-tree && git
read-tree`. But the original shell script only called `git write-tree`
(because the read-tree would obviously be a no-op). So let's skip the
reset_tree() call that is the equivalent that that `git read-tree`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>